Update after feedback
- Repository has been renamed to Factory
- Storable has been renamed to Repository
- Business logic has been moved back to Models
- Factories for all the Entities/Models
