+++ src/DI.php
@@ -28,9 +28,10 @@ use Psr\Log\LoggerInterface;
  * @method static Core\Process process()
  * @method static Core\Session\ISession session()
  * @method static Database\Database dba()
+ * @method static Factory\Mastodon\FollowRequest mstdnFollowRequest()
+ * @method static Factory\Mastodon\Relationship mstdnRelationship()
  * @method static Model\Notify notify()
- * @method static Repository\Mastodon\FollowRequest mstdnFollowRequest()
- * @method static Repository\Storable\Introduction intro()
+ * @method static Repository\Introduction intro()
  * @method static Protocol\Activity activity()
  * @method static Util\ACLFormatter aclFormatter()
  * @method static Util\DateTimeFormat dtFormat()
@@ -62,9 +63,10 @@ abstract class DI
 		'process'      => Core\Process::class,
 		'session'      => Core\Session\ISession::class,
 		'dba'          => Database\Database::class,
+		'mstdnFollowRequest' => Factory\Mastodon\FollowRequest::class,
+		'mstdnRelationship'  => Factory\Mastodon\Relationship::class,
 		'notify'       => Model\Notify::class,
-		'intro'        => Repository\Storable\Introduction::class,
-		'mstdnFollowRequest' => Repository\Mastodon\FollowRequest::class,
+		'intro'        => Repository\Introduction::class,
 		'activity'     => Protocol\Activity::class,
 		'aclFormatter' => Util\ACLFormatter::class,
 		'dtFormat'     => Util\DateTimeFormat::class,
