+++ src/Factory/Mastodon/Account.php
@@ -1,14 +1,14 @@
 <?php
 
-namespace Friendica\Repository\Mastodon;
+namespace Friendica\Factory\Mastodon;
 
 use Friendica\DI;
 use Friendica\Model\APContact;
 use Friendica\Model\Contact;
 use Friendica\Network\HTTPException;
-use Friendica\Repository;
+use Friendica\Factory;
 
-class Account extends Repository
+class Account extends Factory
 {
 	/**
 	 * @param int $contactId
